fix chat unique constraint by buyer seller product and rating

+++ internal/models/comment.go
@@ -8,7 +8,7 @@ type Comment struct {
 	ID        uint `gorm:"primaryKey"`
 	ProductID uint
 	UserID    uint
-	Rating    int
+	Rating    float64
 	Text      string
 	CreatedAt time.Time
 }
